requirements.txt
aiogram
aiohttp
sqlalchemy
aiosqlite
python-dotenv
openai
